+++ src/Facades/LaravelSerializer.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Mochaka\LaravelSerializer\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+ * @see \Mochaka\LaravelSerializer\LaravelSerializer
+ */
+class LaravelSerializer extends Facade
+{
+	protected static function getFacadeAccessor()
+	{
+		return \Mochaka\LaravelSerializer\LaravelSerializer::class;
+	}
+}