+++ src/LaravelSerializerServiceProvider.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Mochaka\LaravelSerializer;
+
+use Spatie\LaravelPackageTools\Package;
+use Spatie\LaravelPackageTools\PackageServiceProvider;
+
+class LaravelSerializerServiceProvider extends PackageServiceProvider
+{
+	public function configurePackage(Package $package): void
+	{
+		/*
+		 * This class is a Package Service Provider
+		 *
+		 * More info: https://github.com/spatie/laravel-package-tools
+		 */
+		$package
+			->name('laravel-serializer')
+			->hasConfigFile();
+	}
+}