+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `laravel-serializer` will be documented in this file.
+
+- - -